# backend/generated-code/src/gen/java/main/com/sqa/thermometer/gen/api/SurveyApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sqa.thermometer.gen.api;

import com.sqa.thermometer.gen.model.SurveyDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-03-24T18:53:38.194613+01:00[Europe/Madrid]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Survey", description = "This will be use to recive user responses")
public interface SurveyApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /survey : Insert a surver
     * Insert a survey data
     *
     * @param surveyDTO Survey to insert in the database (required)
     * @return Successful operation (status code 201)
     */
    @Operation(
        operationId = "postSurvey",
        summary = "Insert a surver",
        description = "Insert a survey data",
        tags = { "Survey" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Successful operation")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/survey",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> postSurvey(
        @Parameter(name = "SurveyDTO", description = "Survey to insert in the database", required = true) @Valid @RequestBody SurveyDTO surveyDTO
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}