# backend/generated-code/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: Api that will measure the happyness of the teams
  title: Team happiness - Buddy program
  version: 1.0.0
servers:
- url: https://localhost:8080
...
paths:
...
  /survey:
    post:
      description: Insert a survey data
      operationId: postSurvey
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Survey'
        description: Survey to insert in the database
        required: true
      responses:
        "201":
          description: Successful operation
      summary: Insert a surver
      tags:
      - Survey
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Survey
...
components:
  schemas:
...
    Survey:
      properties:
        surveyId:
          example: b2f7a764-14b5-45dd-9222-0a4a746f4f7d
          format: UUID
          type: string
        teamId:
          example: b2f7a764-14b5-45dd-9222-0a4a746f4f7d
          format: UUID
          type: string
      required:
      - surveyId
      - teamId
      type: object
...